rotate the y axis in heatmap and move the function to pipeline_auxiliaries
